jifcam_back/requirements.txt
astroid==2.2.3
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-rest-auth==0.9.3
django-uuidfield==0.5.0
djangorestframework==3.9.2
fcm-django==0.2.21
gunicorn==19.9.0
idna==2.8
isort==4.3.10
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
Pillow==5.4.1
psycopg2-binary==2.7.7
pyfcm==1.4.5
pylint==2.3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
six==1.12.0
typed-ast==1.2.0
urllib3==1.24.1
wrapt==1.11.1
